refactor(#421): extract getStatusMeta/statusClass/statusBadgeClass to utils.ts
- Create providers/utils.ts with verification status helpers
- Remove ~55 lines of duplicated logic from ProvidersPage

Partial #421 — EditSheet/ProviderRow extractions to follow.
